+++ lexnlp/extract/all_locales/geoentities.py
@@ -2,10 +2,11 @@
 
 __author__ = "ContraxSuite, LLC; LexPredict, LLC"
 __copyright__ = "Copyright 2015-2021, ContraxSuite, LLC"
-__license__ = "https://github.com/LexPredict/lexpredict-lexnlp/blob/2.2.0/LICENSE"
-__version__ = "2.2.0"
+__license__ = "https://github.com/LexPredict/lexpredict-lexnlp/blob/2.2.1.0/LICENSE"
+__version__ = "2.2.1.0"
 __maintainer__ = "LexPredict, LLC"
 __email__ = "support@contraxsuite.com"
+
 
 from typing import Generator, List, Optional, Tuple, Dict
 
@@ -14,14 +15,6 @@ from lexnlp.extract.en.dict_entities import DictionaryEntry
 from lexnlp.extract.all_locales.languages import LANG_EN, LANG_DE, DEFAULT_LANGUAGE, Locale
 from lexnlp.extract.en.geoentities import get_geoentity_annotations as get_geoentity_annotations_en
 from lexnlp.extract.de.geoentities import get_geoentity_annotations as get_geoentity_annotations_de
-
-
-__author__ = "ContraxSuite, LLC; LexPredict, LLC"
-__copyright__ = "Copyright 2015-2020, ContraxSuite, LLC"
-__license__ = "https://github.com/LexPredict/lexpredict-lexnlp/blob/2.2.0/LICENSE"
-__version__ = "2.2.0"
-__maintainer__ = "LexPredict, LLC"
-__email__ = "support@contraxsuite.com"
 
 
 ROUTINE_BY_LOCALE = {