+++ lexnlp/extract/en/dates.py
@@ -5,10 +5,11 @@ This module implements date extraction functionality in English.
 
 __author__ = "ContraxSuite, LLC; LexPredict, LLC"
 __copyright__ = "Copyright 2015-2021, ContraxSuite, LLC"
-__license__ = "https://github.com/LexPredict/lexpredict-lexnlp/blob/2.2.0/LICENSE"
-__version__ = "2.2.0"
+__license__ = "https://github.com/LexPredict/lexpredict-lexnlp/blob/2.2.1.0/LICENSE"
+__version__ = "2.2.1.0"
 __maintainer__ = "LexPredict, LLC"
 __email__ = "support@contraxsuite.com"
+
 
 # pylint: disable=bare-except
 
@@ -32,6 +33,8 @@ from lexnlp.extract.en.date_model import MODEL_DATE, MODULE_PATH, DATE_MODEL_CHA
 
 
 # Distance in characters to use to merge two date strings
+
+
 DATE_MERGE_WINDOW = 10
 
 # Maximum date length