+++ lexnlp/ml/__init__.py
@@ -1,0 +1,6 @@
+__author__ = "ContraxSuite, LLC; LexPredict, LLC"
+__copyright__ = "Copyright 2015-2021, ContraxSuite, LLC"
+__license__ = "https://github.com/LexPredict/lexpredict-lexnlp/blob/2.2.1.0/LICENSE"
+__version__ = "2.2.1.0"
+__maintainer__ = "LexPredict, LLC"
+__email__ = "support@contraxsuite.com"