+++ lexnlp/utils/amount_delimiting.py
@@ -1,9 +1,10 @@
 __author__ = "ContraxSuite, LLC; LexPredict, LLC"
 __copyright__ = "Copyright 2015-2021, ContraxSuite, LLC"
-__license__ = "https://github.com/LexPredict/lexpredict-lexnlp/blob/2.2.0/LICENSE"
-__version__ = "2.2.0"
+__license__ = "https://github.com/LexPredict/lexpredict-lexnlp/blob/2.2.1.0/LICENSE"
+__version__ = "2.2.1.0"
 __maintainer__ = "LexPredict, LLC"
 __email__ = "support@contraxsuite.com"
+
 
 # -*- coding: utf-8 -*-
 
@@ -13,6 +14,8 @@ from lexnlp.extract.all_locales.languages import LocaleContextManager
 
 
 # https://en.wikipedia.org/wiki/Decimal_separator
+
+
 DELIMITERS: FrozenSet = frozenset((
     '\u0020',  # U+0020   SPACE (HTML &#32;)
     '\u0027',  # U+0027 ' APOSTROPHE (HTML &#39; · &apos;)